@@ -24,8 +24,7 @@
   "scripts": {
     "changelog": "lerna-changelog",
     "lint": "eslint .",
-    "test": "npm-run-all lint test:jest",
-    "test:jest": "jest"
+    "test": "jest"
   },
   "jest": {
     "testMatch": [
@@ -33,30 +32,30 @@
     ]
   },
   "dependencies": {
-    "semver": "^5.3.0"
+    "semver": "^7.6.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "babel-core": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "chai": "^4.2.0",
-    "eslint": "^4.19.1",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-node": "^6.0.1",
-    "eslint-plugin-prettier": "^2.6.0",
-    "jest": "^21.0.0",
-    "lerna-changelog": "^0.8.2",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^1.12.1",
-    "release-it": "^14.2.1",
-    "release-it-lerna-changelog": "^3.1.0"
+    "@babel/core": "^7.24.4",
+    "@babel/preset-env": "^7.24.4",
+    "@eslint/js": "^9.0.0",
+    "@release-it-plugins/lerna-changelog": "^6.1.0",
+    "chai": "^5.1.0",
+    "code-equality-assertions": "^0.9.0",
+    "eslint": "^9.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-prettier": "^5.1.3",
+    "globals": "^15.0.0",
+    "jest": "^29.7.0",
+    "lerna-changelog": "^2.2.0",
+    "prettier": "^3.2.5",
+    "release-it": "^17.1.1"
   },
   "peerDependencies": {
     "@babel/core": "^7.0.0"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=16"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org"
