@@ -1,5 +1,3 @@
-if (true
-/* DEBUG */
-) {
+if (true/* DEBUG */) {
   console.log('woot');
 }
