@@ -1,8 +1,6 @@
 export let PartialComponentManager;
 
-if (false
-/* DEPRECATED_PARTIALS */
-) {
+if (false/* DEPRECATED_PARTIALS */) {
   PartialComponentManager = class {
     constructor() {
       this.isDone = true;
@@ -11,29 +9,21 @@ if (false
   };
 }
 
-if (false
-/* DEPRECATED_PARTIALS */
-&& someOtherThing()) {
+if (false/* DEPRECATED_PARTIALS */&& someOtherThing()) {
   doStuff();
 }
 
-if (!false
-/* DEPRECATED_PARTIALS */
-&& someOtherThing()) {
+if (!false/* DEPRECATED_PARTIALS */&& someOtherThing()) {
   doStuff2();
 }
 
-if (false
-/* DEPRECATED_PARTIALS */
-=== false && someOtherThing()) {
+if (false/* DEPRECATED_PARTIALS */=== false && someOtherThing()) {
   doStuff3();
 }
 
 export let ObjectController;
 
-if (true
-/* DEPRECATED_CONTROLLERS */
-) {
+if (true/* DEPRECATED_CONTROLLERS */) {
   ObjectController = class {
     constructor() {
       this.isDoneAsWell = true;
