@@ -2,8 +2,8 @@
 
 const DebugToolsPlugin = require('..');
 
-describe('baseDir', function() {
-  it('returns directory with package.json', function() {
+describe('baseDir', function () {
+  it('returns directory with package.json', function () {
     let baseDir = DebugToolsPlugin.baseDir();
     let pkg = require(baseDir + '/package');
 
